Refactor headless namespace
