+++ Tests/Unit/Infrastructure/Sulu/Headless/Content/DataProviderResolver/ArticlePageTreeDataProviderResolverTest.php
@@ -11,22 +11,28 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\ArticleBundle\Tests\Unit\Infrastructure\SuluHeadlessBundle\Content\ContentTypeResolver;
+namespace Sulu\Bundle\ArticleBundle\Tests\Unit\Infrastructure\Sulu\Headless\Content\DataProviderResolver;
 
 use PHPUnit\Framework\TestCase;
-use Prophecy\Argument;
 use Prophecy\Prophecy\ObjectProphecy;
-use Sulu\Bundle\ArticleBundle\Infrastructure\SuluHeadlessBundle\ContentTypeResolver\ArticleSelectionResolver;
-use Sulu\Bundle\HeadlessBundle\Content\ContentView;
+use Sulu\Bundle\ArticleBundle\Infrastructure\Sulu\Headless\DataProviderResolver\ArticlePageTreeDataProviderResolver;
 use Sulu\Bundle\HeadlessBundle\Content\StructureResolverInterface;
-use Sulu\Component\Content\Compat\PropertyInterface;
 use Sulu\Component\Content\Compat\PropertyParameter;
 use Sulu\Component\Content\Compat\StructureInterface;
 use Sulu\Component\Content\Mapper\ContentMapperInterface;
 use Sulu\Component\Content\Query\ContentQueryBuilderInterface;
+use Sulu\Component\SmartContent\Configuration\ProviderConfigurationInterface;
+use Sulu\Component\SmartContent\DataProviderInterface;
+use Sulu\Component\SmartContent\DataProviderResult;
+use Sulu\Component\SmartContent\ResourceItemInterface;
 
-class ArticleSelectionResolverTest extends TestCase
+class ArticlePageTreeDataProviderResolverTest extends TestCase
 {
+    /**
+     * @var DataProviderInterface|ObjectProphecy
+     */
+    private $articleDataProvider;
+
     /**
      * @var StructureResolverInterface|ObjectProphecy
      */
@@ -43,17 +49,19 @@ class ArticleSelectionResolverTest extends TestCase
     private $contentMapper;
 
     /**
-     * @var ArticleSelectionResolver
+     * @var ArticlePageTreeDataProviderResolver
      */
-    private $articleSelectionResolver;
+    private $articleDataProviderResolver;
 
     protected function setUp(): void
     {
+        $this->articleDataProvider = $this->prophesize(DataProviderInterface::class);
         $this->structureResolver = $this->prophesize(StructureResolverInterface::class);
         $this->contentQueryBuilder = $this->prophesize(ContentQueryBuilderInterface::class);
         $this->contentMapper = $this->prophesize(ContentMapperInterface::class);
 
-        $this->articleSelectionResolver = new ArticleSelectionResolver(
+        $this->articleDataProviderResolver = new ArticlePageTreeDataProviderResolver(
+            $this->articleDataProvider->reveal(),
             $this->structureResolver->reveal(),
             $this->contentQueryBuilder->reveal(),
             $this->contentMapper->reveal(),
@@ -61,33 +69,59 @@ class ArticleSelectionResolverTest extends TestCase
         );
     }
 
-    public function testGetContentType(): void
+    public function testGetDataProvider(): void
     {
-        self::assertSame('article_selection', $this->articleSelectionResolver::getContentType());
+        self::assertSame('articles_page_tree', $this->articleDataProviderResolver::getDataProvider());
+    }
+
+    public function testGetProviderConfiguration(): void
+    {
+        $configuration = $this->prophesize(ProviderConfigurationInterface::class);
+        $this->articleDataProvider->getConfiguration()->willReturn($configuration->reveal());
+
+        $this->assertSame($configuration->reveal(), $this->articleDataProviderResolver->getProviderConfiguration());
+    }
+
+    public function testGetProviderDefaultParams(): void
+    {
+        $propertyParameter = $this->prophesize(PropertyParameter::class);
+        $this->articleDataProvider->getDefaultPropertyParameter()->willReturn(['test' => $propertyParameter->reveal()]);
+
+        $this->assertSame(['test' => $propertyParameter->reveal()], $this->articleDataProviderResolver->getProviderDefaultParams());
     }
 
     public function testResolve(): void
     {
-        $structure = $this->prophesize(StructureInterface::class);
-        $structure->getWebspaceKey()->willReturn('webspace-key');
+        $providerResultItem1 = $this->prophesize(ResourceItemInterface::class);
+        $providerResultItem1->getId()->willReturn('article-id-1');
 
-        /** @var PropertyInterface|ObjectProphecy $property */
-        $property = $this->prophesize(PropertyInterface::class);
-        $params = [
+        $providerResultItem2 = $this->prophesize(ResourceItemInterface::class);
+        $providerResultItem2->getId()->willReturn('article-id-2');
+
+        $providerResult = $this->prophesize(DataProviderResult::class);
+        $providerResult->getHasNextPage()->willReturn(true);
+        $providerResult->getItems()->willReturn([$providerResultItem1->reveal(), $providerResultItem2->reveal()]);
+
+        $propertyParameters = [
             'properties' => new PropertyParameter('properties', [
                 new PropertyParameter('contentDescription', 'description'),
                 new PropertyParameter('excerptTitle', 'excerpt.title'),
-                new PropertyParameter('categories', 'excerpt.categories'),
             ]),
         ];
 
-        $property->getParams()->willReturn($params);
-        $property->getStructure()->willReturn($structure->reveal());
-
         // expected and unexpected service calls
+        $this->articleDataProvider->resolveResourceItems(
+            ['filter-key' => 'filter-value'],
+            $propertyParameters,
+            ['webspaceKey' => 'webspace-key', 'locale' => 'en'],
+            10,
+            1,
+            5
+        )->willReturn($providerResult->reveal())->shouldBeCalledOnce();
+
         $this->contentQueryBuilder->init([
             'ids' => ['article-id-1', 'article-id-2'],
-            'properties' => $params['properties']->getValue(),
+            'properties' => $propertyParameters['properties']->getValue(),
             'published' => false,
         ])->shouldBeCalled();
         $this->contentQueryBuilder->build('webspace-key', ['en'])->willReturn(['article-query-string']);
@@ -96,7 +130,6 @@ class ArticleSelectionResolverTest extends TestCase
         $articleStructure1->getUuid()->willReturn('article-id-1');
         $articleStructure2 = $this->prophesize(StructureInterface::class);
         $articleStructure2->getUuid()->willReturn('article-id-2');
-
         $this->contentMapper->loadBySql2(
             'article-query-string',
             'en',
@@ -105,6 +138,7 @@ class ArticleSelectionResolverTest extends TestCase
             $articleStructure2->reveal(),
             $articleStructure1->reveal(),
         ])->shouldBeCalledOnce();
+
         $this->structureResolver->resolveProperties(
             $articleStructure1->reveal(),
             [
@@ -112,7 +146,6 @@ class ArticleSelectionResolverTest extends TestCase
                 'routePath' => 'routePath',
                 'contentDescription' => 'description',
                 'excerptTitle' => 'excerpt.title',
-                'categories' => 'excerpt.categories',
             ],
             'en'
         )->willReturn([
@@ -123,14 +156,12 @@ class ArticleSelectionResolverTest extends TestCase
                 'routePath' => '/article-url-1',
                 'contentDescription' => 'Article Content Description',
                 'excerptTitle' => 'Article Excerpt Title 1',
-                'categories' => [],
             ],
             'view' => [
                 'title' => [],
                 'routePath' => [],
                 'contentDescription' => [],
                 'excerptTitle' => [],
-                'categories' => [],
             ],
         ])->shouldBeCalledOnce();
 
@@ -141,7 +172,6 @@ class ArticleSelectionResolverTest extends TestCase
                 'routePath' => 'routePath',
                 'contentDescription' => 'description',
                 'excerptTitle' => 'excerpt.title',
-                'categories' => 'excerpt.categories',
             ],
             'en'
         )->willReturn([
@@ -152,25 +182,26 @@ class ArticleSelectionResolverTest extends TestCase
                 'routePath' => '/article-url-2',
                 'contentDescription' => 'Article Content Description',
                 'excerptTitle' => 'Article Excerpt Title 2',
-                'categories' => [],
             ],
             'view' => [
                 'title' => [],
                 'routePath' => [],
                 'contentDescription' => [],
                 'excerptTitle' => [],
-                'categories' => [],
             ],
         ])->shouldBeCalledOnce();
 
         // call test function
-        $result = $this->articleSelectionResolver->resolve(
-            ['article-id-1', 'article-id-2'],
-            $property->reveal(),
-            'en'
+        $result = $this->articleDataProviderResolver->resolve(
+            ['filter-key' => 'filter-value'],
+            $propertyParameters,
+            ['webspaceKey' => 'webspace-key', 'locale' => 'en'],
+            10,
+            1,
+            5
         );
 
-        $this->assertInstanceOf(ContentView::class, $result);
+        $this->assertTrue($result->getHasNextPage());
         $this->assertSame(
             [
                 [
@@ -181,14 +212,12 @@ class ArticleSelectionResolverTest extends TestCase
                         'routePath' => '/article-url-1',
                         'contentDescription' => 'Article Content Description',
                         'excerptTitle' => 'Article Excerpt Title 1',
-                        'categories' => [],
                     ],
                     'view' => [
                         'title' => [],
                         'routePath' => [],
                         'contentDescription' => [],
                         'excerptTitle' => [],
-                        'categories' => [],
                     ],
                 ],
                 [
@@ -199,63 +228,57 @@ class ArticleSelectionResolverTest extends TestCase
                         'routePath' => '/article-url-2',
                         'contentDescription' => 'Article Content Description',
                         'excerptTitle' => 'Article Excerpt Title 2',
-                        'categories' => [],
                     ],
                     'view' => [
                         'title' => [],
                         'routePath' => [],
                         'contentDescription' => [],
                         'excerptTitle' => [],
-                        'categories' => [],
                     ],
                 ],
             ],
-            $result->getContent()
+            $result->getItems()
+        );
+    }
+
+    public function testResolveEmptyProviderResult(): void
+    {
+        $providerResult = $this->prophesize(DataProviderResult::class);
+        $providerResult->getHasNextPage()->willReturn(false);
+        $providerResult->getItems()->willReturn([]);
+
+        $propertyParameters = [
+            'properties' => new PropertyParameter('properties', [
+                new PropertyParameter('contentDescription', 'description'),
+                new PropertyParameter('excerptTitle', 'excerpt.title'),
+            ]),
+        ];
+
+        // expected and unexpected service calls
+        $this->articleDataProvider->resolveResourceItems(
+            ['filter-key' => 'filter-value'],
+            $propertyParameters,
+            ['webspaceKey' => 'webspace-key', 'locale' => 'en'],
+            10,
+            1,
+            5
+        )->willReturn($providerResult->reveal())
+            ->shouldBeCalledOnce();
+
+        // call test function
+        $result = $this->articleDataProviderResolver->resolve(
+            ['filter-key' => 'filter-value'],
+            $propertyParameters,
+            ['webspaceKey' => 'webspace-key', 'locale' => 'en'],
+            10,
+            1,
+            5
         );
 
+        $this->assertFalse($result->getHasNextPage());
         $this->assertSame(
-            ['ids' => ['article-id-1', 'article-id-2']],
-            $result->getView()
+            [],
+            $result->getItems()
         );
-    }
-
-    public function testResolveDataIsNull(): void
-    {
-        $locale = 'en';
-        $property = $this->prophesize(PropertyInterface::class);
-
-        // expected and unexpected service calls
-        $this->contentQueryBuilder->init(Argument::cetera())
-            ->shouldNotBeCalled();
-
-        $this->structureResolver->resolve(Argument::cetera())
-            ->shouldNotBeCalled();
-
-        // call test function
-        $result = $this->articleSelectionResolver->resolve(null, $property->reveal(), $locale);
-
-        $this->assertSame([], $result->getContent());
-
-        $this->assertSame(['ids' => []], $result->getView());
-    }
-
-    public function testResolveDataIsEmptyArray(): void
-    {
-        $locale = 'en';
-        $property = $this->prophesize(PropertyInterface::class);
-
-        // expected and unexpected service calls
-        $this->contentQueryBuilder->init(Argument::any())
-            ->shouldNotBeCalled();
-
-        $this->structureResolver->resolve(Argument::any())
-            ->shouldNotBeCalled();
-
-        // call test function
-        $result = $this->articleSelectionResolver->resolve([], $property->reveal(), $locale);
-
-        $this->assertSame([], $result->getContent());
-
-        $this->assertSame(['ids' => []], $result->getView());
     }
 }
